Implemented a rolling enemy.

+++ Assets/CharacterStats.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "CharacterStats", menuName = "Scriptable Objects/CharacterStats")]
 public class CharacterStats : ScriptableObject
 {
-    public float maxHealth;
-    public float attackPower;
+    public int maxHealth;
+    public int attackPower;
     public float movementSpeed;
 }
